BREAKING: Modify path layout to fix relative RPATH lookups (#104)
The executable of an executable transitioned rule is now created in a
subdirectory of the target's package, which allows it to match the path
depth under the execroot of the original executable and thus gets
relative RPATH entries to resolve.

Also copy any DLLs next to the new executable on Windows.

The new path layout can be disabled via
`--no@with_cfg.bzl//:incompatible_same_depth_path_layout`, e.g. if users
hardcoded rlocationpaths and don't want to update them now. It is
recommended to avoid this and use `$(rlocationpath ...)` instead.

Fixes #85