Use `rule(parent = ...)` when possible (#145)
When applying `with_cfg` to a rule with Bazel 8, it can be extended with
a transition rather than being wrapped, which avoids a lot of dirty
tricks needed to forward common attributes and providers.

This doesn't work with automatic reset yet as it would require returning
two rules and a macro. In the future, this could be resolved by merging
the transitioning alias into the extended rule and returning the macro
first.