Catch missing `Label(...)` usage in the transition
When users transition a flag such as `--platforms` to a list of strings rather than a list of `Label`s, the label strings will resolve relative to the `with_cfg.bzl` repo, which is never correct. Instead of showing an unhelpful error in that case, catch this type of issue early and show a more helpful message.

Also add an example of a transition on `--platforms` that can be used to test the new check by dropping the `Label` call.